removed SRE console access of CCS project + refactoring
