+++ pkg/controller/projectreference/projectreference_adapter.go
@@ -124,6 +124,7 @@ func NewReferenceAdapter(
 	return r, nil
 }
 
+// EnsureProjectClaimReady sets the ProjectClaim to Ready after the ProjectReference was reconciled correctly and gcp project has been created
 func EnsureProjectClaimReady(r *ReferenceAdapter) (gcputil.OperationResult, error) {
 	if r.ProjectReference.Status.State != gcpv1alpha1.ProjectReferenceStatusReady {
 		return gcputil.ContinueProcessing()
@@ -152,6 +153,7 @@ func EnsureProjectClaimReady(r *ReferenceAdapter) (gcputil.OperationResult, erro
 	return gcputil.StopProcessing()
 }
 
+// Waits until the ProjectClaim has been initialized, meaning is in state PendingProject
 func VerifyProjectClaimPending(r *ReferenceAdapter) (gcputil.OperationResult, error) {
 	if r.ProjectClaim.Status.State != gcpv1alpha1.ClaimStatusPendingProject {
 		return gcputil.RequeueAfter(5*time.Second, nil)
@@ -375,6 +377,18 @@ func (r *ReferenceAdapter) EnsureProjectCleanedUp() error {
 		err = r.deleteProject()
 		if err != nil {
 			return err
+		}
+	} else {
+		r.logger.Info("Deleting IAM Policy for console access")
+		for _, email := range r.OperatorConfig.CCSConsoleAccess {
+			if err := r.DeleteIAMPolicy(email, gcputil.GoogleGroup); err != nil {
+				return err
+			}
+		}
+		for _, email := range r.OperatorConfig.CCSReadOnlyConsoleAccess {
+			if err := r.DeleteIAMPolicy(email, gcputil.GoogleGroup); err != nil {
+				return err
+			}
 		}
 	}
 
@@ -702,7 +716,6 @@ func EnsureProjectReferenceInitialized(r *ReferenceAdapter) (gcputil.OperationRe
 		if err != nil {
 			return gcputil.RequeueWithError(operrors.Wrap(err, "failed to initalize ProjectReference"))
 		}
-		return gcputil.StopProcessing()
 	}
 	return gcputil.ContinueProcessing()
 }
@@ -735,34 +748,65 @@ func (r *ReferenceAdapter) AddOrUpdateBindings(serviceAccountEmail string, polic
 func (r *ReferenceAdapter) SetIAMPolicy(serviceAccountEmail string, policies []string, memberType gcputil.IamMemberType) error {
 	// Checking if policy needs to be updated
 	var retry int
+	addorUpdateResponse, err := r.AddOrUpdateBindings(serviceAccountEmail, policies, memberType)
+	if err != nil {
+		return err
+	}
+	if !addorUpdateResponse.modified {
+		return nil
+	}
+	setIamPolicyRequest := &cloudresourcemanager.SetIamPolicyRequest{
+		Policy: addorUpdateResponse.policy,
+	}
+
 	for {
-		retry++
-		time.Sleep(time.Second)
-		addorUpdateResponse, err := r.AddOrUpdateBindings(serviceAccountEmail, policies, memberType)
+		_, err = r.gcpClient.SetIamPolicy(setIamPolicyRequest)
 		if err != nil {
+			ae, ok := err.(*googleapi.Error)
+			// retry rules below:
+			if ok && ae.Code == http.StatusConflict && retry <= 3 {
+				retry++
+				time.Sleep(time.Second)
+				continue
+			}
 			return err
-		}
-
-		// If existing bindings have been modified update the policy
-		if addorUpdateResponse.modified {
-			setIamPolicyRequest := &cloudresourcemanager.SetIamPolicyRequest{
-				Policy: addorUpdateResponse.policy,
-			}
-			_, err = r.gcpClient.SetIamPolicy(setIamPolicyRequest)
-			if err != nil {
-				ae, ok := err.(*googleapi.Error)
-				// retry rules below:
-
-				if ok && ae.Code == http.StatusConflict && retry <= 3 {
-					continue
-				}
-				return err
-			}
-			return nil
 		}
 		return nil
 	}
+}
 
+func (r *ReferenceAdapter) DeleteIAMPolicy(serviceAccountEmail string, memberType gcputil.IamMemberType) error {
+	var retry int
+	policies, err := r.gcpClient.GetIamPolicy(r.ProjectReference.Spec.GCPProjectID)
+	if err != nil {
+		return err
+	}
+	for i, binding := range policies.Bindings {
+		for index, v := range binding.Members {
+			if v == gcputil.PrefixMemberType[memberType]+serviceAccountEmail {
+				newMembers := append(binding.Members[:index], binding.Members[index+1:]...)
+				policies.Bindings[i].Members = newMembers
+				break
+			}
+		}
+	}
+	setIamPolicyRequest := &cloudresourcemanager.SetIamPolicyRequest{
+		Policy: policies,
+	}
+	for {
+		_, err = r.gcpClient.SetIamPolicy(setIamPolicyRequest)
+		if err != nil {
+			ae, ok := err.(*googleapi.Error)
+			// retry rules below:
+			if ok && ae.Code == http.StatusConflict && retry <= 3 {
+				retry++
+				time.Sleep(time.Second)
+				continue
+			}
+			return err
+		}
+		return nil
+	}
 }
 
 // SetProjectReferenceCondition calls SetCondition() with project reference conditions
@@ -790,7 +834,7 @@ func (r *ReferenceAdapter) SetProjectReferenceCondition(reason string, err error
 func (r *ReferenceAdapter) StatusUpdate() error {
 	err := r.kubeClient.Status().Update(context.TODO(), r.ProjectReference)
 	if err != nil {
-		return operrors.Wrap(err, fmt.Sprintf("failed to update ProjectClaim state for %s", r.ProjectReference.Name))
+		return operrors.Wrap(err, fmt.Sprintf("failed to update ProjectReference status of %s", r.ProjectReference.Name))
 	}
 
 	return nil
